Make sure that modules are abandoned if TP tells Souvenir to (when autosolving).
In the previous code, the sequence would be: yield return (which is usually the WaitForSeconds waiting for the module to be solved), then MoveNext() (which can add the questions if the module is now solved), and THEN check TwitchAbandonModule, which is too late now.
New code: yield return, then check TwitchAbandonModule; this way, if Twitch added the module during the wait, it no longer gets a chance to add questions.