Module Maze: remove gSprites array
Module Maze - all sprites are now available in the sprites array, and gSprites has been removed. The gSprites (grayscaleSprites) array was created for Souvenir support, back when the icons on the screen were intended to be in grayscale as opposed to in color. This feature was removed soon after its implementation, as such there is no need for two separate sprite arrays.