Add the option to provide custom formatted module names
formattedModuleName is a new optional parameter in the makeQuestion and AddQuestion methods. If set, its value will be used instead of the default formatting based on the solve count.
Also make it so the module will wait for all supported modules on the bomb to solve (other than itself and other Souvenir modules), as this is helpful for supported boss modules.